Apply authorization

+++ resources/views/student/detail.blade.php
@@ -125,14 +125,19 @@
         <div class="col-xs-12 col-md-4 col-md-offset-4">
             <canvas id="myChart" style="background-color: white" width="200" height="200"></canvas>
         </div>
-        @if (Auth::check())
+        @if (Auth::check() && Auth::user()->role != 1)
             <div class="col-xs-12" style="height:50px;"></div>
-            {!! Form::open(['method' => 'DELETE', 'onsubmit' => 'return ConfirmDelete()']) !!}
-            <div class="form-group col-md-6 col-lg-4 col-lg-offset-2">
-                {!! Form::hidden('id', $student->id) !!}
-                {!! Form::submit('Delete', ['class' => 'form-control btn-danger']) !!}
-            </div>
-            <a class="col-md-6 col-lg-4 btn btn-success" href=<?php echo '"/student/' . $student->id . '/edit"' ?>>Update</a>
+            @if (Auth::user()->role == 2)
+                {!! Form::open(['method' => 'DELETE', 'onsubmit' => 'return ConfirmDelete()']) !!}
+                <div class="form-group col-md-6 col-lg-4 col-lg-offset-2">
+                    {!! Form::hidden('id', $student->id) !!}
+                    {!! Form::submit('Delete', ['class' => 'form-control btn-danger']) !!}
+                </div>
+                {!! Form::close() !!}
+                <a class="col-md-6 col-lg-4 btn btn-success" href=<?php echo '"/student/' . $student->id . '/edit"' ?>>Update</a>
+            @else
+                <a class="col-md-8 col-lg-4 col-md-offset-2 col-lg-offset-4 btn btn-success" href=<?php echo '"/student/' . $student->id . '/edit"' ?>>Update</a>
+            @endif
         @endif
     </div>
